moved AreTheSameAxes functions to Rewrite.cpp
since they are only used in Rewrite.td
