+++ src/Dialect/ONNX/ONNXOps/OpHelper.hpp
@@ -182,15 +182,8 @@ mlir::ArrayAttr CombinedTransposePattern(mlir::PatternRewriter &rewriter,
 /// Identity patterns are {0, 1, 2, ... , rank -1}.
 bool IsIdentityPermuteVector(mlir::ArrayAttr permAttr);
 
-/// Test if two axis arrays contain the same values or not.
-bool AreTheSameAxesArrayAttr(
-    int64_t rank, mlir::ArrayAttr lhsAttr, mlir::ArrayAttr rhsAttr);
-
 /// Test if the value has the specified constant shape
 bool HasSpecifiedConstantShape(mlir::Value value, mlir::Value shape);
-
-/// Test if two values are constants with the same axis values.
-bool AreTheSameAxesConstant(int64_t rank, mlir::Value lhs, mlir::Value rhs);
 
 /// Test if 'val' has shape and rank or not.
 bool hasShapeAndRank(mlir::Value val);
